+++ packages/manager/src/features/Domains/CreateDomain/CreateDomain.tsx
@@ -1,6 +1,5 @@
 import {
   CreateDomainPayload,
-  createDomainRecord,
   Domain,
   DomainType,
 } from '@linode/api-v4/lib/domains';
@@ -39,12 +38,14 @@ import {
   handleFieldErrors,
   handleGeneralErrors,
 } from 'src/utilities/formikErrorUtils';
+import { sendCreateDomainEvent } from 'src/utilities/ga';
 import {
   ExtendedIP,
   extendedIPToString,
   stringToExtendedIP,
 } from 'src/utilities/ipUtils';
 import scrollErrorIntoView from 'src/utilities/scrollErrorIntoView';
+import { generateDefaultDomainRecords } from '../domainUtils';
 
 const useStyles = makeStyles((theme: Theme) => ({
   main: {
@@ -75,69 +76,6 @@ const useStyles = makeStyles((theme: Theme) => ({
 }));
 
 type DefaultRecordsType = 'none' | 'linode' | 'nodebalancer';
-
-export const generateDefaultDomainRecords = (
-  domain: string,
-  domainID: number,
-  ipv4?: string,
-  ipv6?: string | null
-) => {
-  /**
-   * At this point, the IPv6 is including the prefix and we need to strip that
-   *
-   * BUT
-   *
-   * this logic only applies to Linodes' ipv6, not NodeBalancers. No stripping
-   * needed for NodeBalancers.
-   */
-  const cleanedIPv6 =
-    ipv6 && ipv6.includes('/') ? ipv6.substr(0, ipv6.indexOf('/')) : ipv6;
-
-  const baseIPv4Requests = [
-    createDomainRecord(domainID, {
-      type: 'A',
-      target: ipv4,
-    }),
-    createDomainRecord(domainID, {
-      type: 'A',
-      target: ipv4,
-      name: 'www',
-    }),
-    createDomainRecord(domainID, {
-      type: 'A',
-      target: ipv4,
-      name: 'mail',
-    }),
-  ];
-
-  return Promise.all(
-    /** ipv6 can be null so don't try to create domain records in that case */
-    !!cleanedIPv6
-      ? [
-          ...baseIPv4Requests,
-          createDomainRecord(domainID, {
-            type: 'AAAA',
-            target: cleanedIPv6,
-          }),
-          createDomainRecord(domainID, {
-            type: 'AAAA',
-            target: cleanedIPv6,
-            name: 'www',
-          }),
-          createDomainRecord(domainID, {
-            type: 'AAAA',
-            target: cleanedIPv6,
-            name: 'mail',
-          }),
-          createDomainRecord(domainID, {
-            type: 'MX',
-            priority: 10,
-            target: `mail.${domain}`,
-          }),
-        ]
-      : baseIPv4Requests
-  );
-};
 
 export const CreateDomain = () => {
   const classes = useStyles();
@@ -258,6 +196,7 @@ export const CreateDomain = () => {
         if (!mounted) {
           return;
         }
+        sendCreateDomainEvent('Domain Create Page');
         /**
          * Now we check to see if the user wanted us to automatically create
          * domain records for them. If so, create some A/AAAA and MX records
