+++ packages/manager/src/features/Domains/domainUtils.ts
@@ -1,4 +1,8 @@
-import { DomainRecord, DomainType } from '@linode/api-v4/lib/domains';
+import {
+  createDomainRecord,
+  DomainRecord,
+  DomainType,
+} from '@linode/api-v4/lib/domains';
 
 export const isValidDomainRecord = (
   hostname: string,
@@ -45,3 +49,66 @@ export const isEditableNameServer = (nameServerId: number) => {
  */
 export const getDomainDisplayType = (domainType: DomainType) =>
   domainType === 'master' ? 'primary' : 'secondary';
+
+export const generateDefaultDomainRecords = (
+  domain: string,
+  domainID: number,
+  ipv4?: string,
+  ipv6?: string | null
+) => {
+  /**
+   * At this point, the IPv6 is including the prefix and we need to strip that
+   *
+   * BUT
+   *
+   * this logic only applies to Linodes' ipv6, not NodeBalancers. No stripping
+   * needed for NodeBalancers.
+   */
+  const cleanedIPv6 =
+    ipv6 && ipv6.includes('/') ? ipv6.substr(0, ipv6.indexOf('/')) : ipv6;
+
+  const baseIPv4Requests = [
+    createDomainRecord(domainID, {
+      type: 'A',
+      target: ipv4,
+    }),
+    createDomainRecord(domainID, {
+      type: 'A',
+      target: ipv4,
+      name: 'www',
+    }),
+    createDomainRecord(domainID, {
+      type: 'A',
+      target: ipv4,
+      name: 'mail',
+    }),
+  ];
+
+  return Promise.all(
+    /** ipv6 can be null so don't try to create domain records in that case */
+    !!cleanedIPv6
+      ? [
+          ...baseIPv4Requests,
+          createDomainRecord(domainID, {
+            type: 'AAAA',
+            target: cleanedIPv6,
+          }),
+          createDomainRecord(domainID, {
+            type: 'AAAA',
+            target: cleanedIPv6,
+            name: 'www',
+          }),
+          createDomainRecord(domainID, {
+            type: 'AAAA',
+            target: cleanedIPv6,
+            name: 'mail',
+          }),
+          createDomainRecord(domainID, {
+            type: 'MX',
+            priority: 10,
+            target: `mail.${domain}`,
+          }),
+        ]
+      : baseIPv4Requests
+  );
+};
